テストケースの修正　p指定子の引数変更

+++ ft_format_handler.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_format_handler.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sosmiyat <sosmiyat@student.42.fr>          +#+  +:+       +#+        */
+/*   By: miyatasoujo <miyatasoujo@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/24 15:13:41 by sosmiyat          #+#    #+#             */
-/*   Updated: 2024/11/24 17:47:31 by sosmiyat         ###   ########.fr       */
+/*   Updated: 2024/11/30 16:38:06 by miyatasoujo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,6 +32,6 @@ int	ft_format_handler(const char *fmt, va_list ap, int count)
 	else if (*fmt == '%')
 		ft_put_per(&count);
 	else if (*fmt == 'p')
-		ft_put_ptr((unsigned long long)va_arg(ap, void *), &count);
+		ft_put_ptr((uintptr_t)va_arg(ap, void *), &count);
 	return (count);
 }
