+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sosmiyat <sosmiyat@student.42.fr>          +#+  +:+       +#+        */
+/*   By: miyatasoujo <miyatasoujo@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/27 15:35:33 by sosmiyat          #+#    #+#             */
-/*   Updated: 2024/11/24 17:44:20 by sosmiyat         ###   ########.fr       */
+/*   Updated: 2024/11/30 16:39:31 by miyatasoujo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,7 @@
 # include <stdio.h>
 # include <stdlib.h>
 # include <unistd.h>
+# include <inttypes.h>
 
 int		ft_printf(const char *fmt, ...);
 
@@ -25,7 +26,7 @@ int		ft_format_handler(const char *fmt, va_list ap, int count);
 
 void	ft_put_pointer_char(char *s, int *count);
 
-void	ft_put_ptr(unsigned long long i, int *count);
+void	ft_put_ptr(uintptr_t i, int *count);
 
 void	ft_put_hex_upper(unsigned int i, int *count);
 
@@ -39,6 +40,6 @@ void	ft_putchar(char c);
 
 void	ft_put_per(int *count);
 
-void	ft_put_hex_long(unsigned long long i, int *count);
+void	ft_put_hex_long(uintptr_t i, int *count);
 
 #endif
