+++ ft_put_ptr.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_put_ptr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sosmiyat <sosmiyat@student.42.fr>          +#+  +:+       +#+        */
+/*   By: miyatasoujo <miyatasoujo@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/24 17:45:07 by sosmiyat          #+#    #+#             */
-/*   Updated: 2024/11/24 17:46:45 by sosmiyat         ###   ########.fr       */
+/*   Updated: 2024/11/30 16:38:47 by miyatasoujo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_put_ptr(unsigned long long i, int *count)
+void	ft_put_ptr(uintptr_t i, int *count)
 {
 	if (!i)
 	{
